requirements.txt
Django
djangorestframework
psycopg2-binary
drf-yasg
faker
openpyxl
